+++ frontend/helpers/cleanData.tsx
@@ -1,30 +1,68 @@
-import { DataProps, NewEvent, RawEvent } from "@/assets/interfaces";
-import getTime from "./getTime";
+import { DataProps, EventListingProps, KeyProps, NewEvent } from '@/assets/interfaces';
+import getTime from './getTime';
 import trueOrFalse from "./trueOrFalse";
 
-export default function cleanData (eventData: RawEvent[]) {
-  var data: DataProps = {
-        eventData: [],
-        filteredEvents: {
-          ageRequirement: {},
-          cost: {},
-          duration: {},
-          endDates: {},
-          endTimes: {},
-          eventTypes: {},
-          experienceType: {},
-          gameSystems: {},
-          groups: {},
-          locations: {},
-          materialsRequired: [],
-          noTickets: [],
-          startDates: {},
-          startTimes: {},
-          tournaments: []
-        }
-      }
+const getKeyByValue = (object: any, value: string) => {
+  const foundKey = Object.keys(object).find((key) => object[key] === value);
+  return foundKey;
+}
 
-  eventData.map((event, index) => {
+export const cleanData = ({ keyList, eventList }: {
+  keyList: KeyProps;
+  eventList: EventListingProps[];
+}) => {
+  var data: DataProps = {
+    eventData: [],
+    filteredEvents: {
+      ageRequirement: {},
+      cost: {},
+      duration: {},
+      endDates: {},
+      endTimes: {},
+      eventTypes: {},
+      experienceType: {},
+      gameSystems: {},
+      groups: {},
+      locations: {},
+      materialsRequired: [],
+      noTickets: [],
+      startDates: {},
+      startTimes: {},
+      tournaments: []
+    }
+  }
+
+  // Event Times
+  const startDateTimeKey = getKeyByValue(keyList, "Start Date & Time");
+  const endDateTimeKey = getKeyByValue(keyList, "End Date & Time");
+
+  // Event Info
+  const ageRequirementKey = getKeyByValue(keyList, "Age Required");
+  const contactKey = getKeyByValue(keyList, "Email");
+  const costKey = getKeyByValue(keyList, "Cost $");
+  const durationKey = getKeyByValue(keyList, "Duration");
+  const descriptionShortKey = getKeyByValue(keyList, "Short Description");
+  const descriptionLongKey = getKeyByValue(keyList, "Long Description");
+  const eventTypeKey = getKeyByValue(keyList, "Event Type");
+  const expKey = getKeyByValue(keyList, "Experience Required");
+  const gameIdKey = getKeyByValue(keyList, "Game ID");
+  const gameSystemKey = getKeyByValue(keyList, "Game System");
+  const gmNamesKey = getKeyByValue(keyList, "GM Names");
+  const groupKey = getKeyByValue(keyList, "Group");
+  const locationKey = getKeyByValue(keyList, "Location");
+  const materialsKey = getKeyByValue(keyList, "Materials Required Details");
+  const playersMinKey = getKeyByValue(keyList, "Minimum Players");
+  const playersMaxKey = getKeyByValue(keyList, "Maximum Players");
+  const tableNumKey = getKeyByValue(keyList, "Table Number");
+  const ticketCountKey = getKeyByValue(keyList, "Tickets Available");
+  const titleKey = getKeyByValue(keyList, "Title");
+  const tournamentKey = getKeyByValue(keyList, "Tournament?");
+  const roomKey = getKeyByValue(keyList, "Room Name");
+  const roundKey = getKeyByValue(keyList, "Round Number");
+  const roundTotalKey = getKeyByValue(keyList, "Total Rounds");
+  const websiteKey = getKeyByValue(keyList, "Website");
+
+  eventList.map((event: any, index) => {
     var newEvent: NewEvent = {
       id: 0,
       ageRequirement: '',
@@ -57,45 +95,71 @@ export default function cleanData (eventData: RawEvent[]) {
       website: ''
     }
 
-    var rawStart = new Date(event["Start Date & Time"]),
-        rawEnd = new Date(event["End Date & Time"]),
-        ageReq = event['Age Required'],
-        contact = event.Email,
-        cost = Number(event['Cost $']),
-        duration = Number(event.Duration),
-        descriptionShort = event['Short Description'],
-        descriptionLong = event['Long Description'],
+    var rawStart = new Date(event[startDateTimeKey!]),
+        rawEnd = new Date(event[endDateTimeKey!]),
         eventStartDate = rawStart.toLocaleDateString(),
         eventStartTime = getTime(rawStart),
         eventEndDate = rawEnd.toLocaleDateString(),
         eventEndTime = getTime(rawEnd),
-        eventType = event['Event Type'],
-        experienceReq = event['Experience Required'],
-        gameId = event['Game ID'],
-        gameSystem = event['Game System'],
-        gmNames = event['GM Names'],
-        group = event['Group'],
-        location = event['Location']?.toUpperCase(),
-        materials = event['Materials Required Details'],
-        playersMin = Number(event['Minimum Players']),
-        playersMax = Number(event['Maximum Players']),
-        tableNum = Number(event['Table Number']),
-        ticketsAvailable = Number(event['Tickets Available']),
-        title = event['Title'],
-        tournament = trueOrFalse(event['Tournament?']),
-        room = event['Room Name'],
-        round = Number(event['Round Number']),
-        roundTotal = Number(event['Total Rounds']),
-        website = event['Website'];
+        ageReq = event[ageRequirementKey!],
+        contact = event[contactKey!],
+        cost = Number(event[costKey!]),
+        duration = Number(event[durationKey!]),
+        descriptionShort = event[descriptionShortKey!],
+        descriptionLong = event[descriptionLongKey!],
+        eventType = event[eventTypeKey!],
+        experienceReq = event[expKey!],
+        gameId = event[gameIdKey!],
+        gameSystem = event[gameSystemKey!],
+        gmNames = event[gmNamesKey!],
+        group = event[groupKey!],
+        location = event[locationKey!],
+        materials = event[materialsKey!],
+        playersMin = Number(event[playersMinKey!]),
+        playersMax = Number(event[playersMaxKey!]),
+        tableNum = Number(event[tableNumKey!]),
+        ticketsAvailable = Number(event[ticketCountKey!]),
+        title = event[titleKey!],
+        tournament = trueOrFalse(event[tournamentKey!]),
+        room = event[roomKey!],
+        round = Number(event[roundKey!]),
+        roundTotal = Number(event[roundTotalKey!]),
+        website = event[websiteKey!];
+    const newId = index - 1;
 
-    newEvent.id = index
+    newEvent.id = newId;
+
+    // Event start and end dates and times
+    if (!data.filteredEvents.endDates[eventEndDate]) {
+      data.filteredEvents.endDates[eventEndDate] = [];
+    };
+    data.filteredEvents.endDates[eventEndDate].push(newId);
+    newEvent.endDate = eventEndDate;
+
+    if (!data.filteredEvents.endTimes[eventEndTime]) {
+      data.filteredEvents.endTimes[eventEndTime] = [];
+    };
+    data.filteredEvents.endTimes[eventEndTime].push(newId);
+    newEvent.endTime = eventEndTime;
+
+    if (!data.filteredEvents.startDates[eventStartDate]) {
+      data.filteredEvents.startDates[eventStartDate] = [];
+    };
+    data.filteredEvents.startDates[eventStartDate].push(newId);
+    newEvent.startDate = eventStartDate;
+
+    if (!data.filteredEvents.startTimes[eventStartTime]) {
+      data.filteredEvents.startTimes[eventStartTime] = [];
+    };
+    data.filteredEvents.startTimes[eventStartTime].push(newId);
+    newEvent.startTime = eventStartTime;
 
     // Age Requirement
-    newEvent.ageRequirement = ageReq
     if (!data.filteredEvents.ageRequirement[ageReq]) {
       data.filteredEvents.ageRequirement[ageReq] = [];
     }
-    data.filteredEvents.ageRequirement[ageReq].push(index)
+    data.filteredEvents.ageRequirement[ageReq].push(newId)
+    newEvent.ageRequirement = ageReq
 
     // Contact
     if (contact) {
@@ -110,7 +174,7 @@ export default function cleanData (eventData: RawEvent[]) {
     if (!data.filteredEvents.cost[cost]) {
       data.filteredEvents.cost[cost] = []
     }
-    data.filteredEvents.cost[cost].push(index)
+    data.filteredEvents.cost[cost].push(newId)
     newEvent.cost = cost
 
     // Duration
@@ -118,7 +182,7 @@ export default function cleanData (eventData: RawEvent[]) {
       if (!data.filteredEvents.duration[duration]) {
         data.filteredEvents.duration[duration] = []
       }
-      data.filteredEvents.duration[duration].push(index)
+      data.filteredEvents.duration[duration].push(newId)
     }
     newEvent.duration = duration
 
@@ -126,54 +190,35 @@ export default function cleanData (eventData: RawEvent[]) {
     newEvent.descriptionShort = descriptionShort
     newEvent.descriptionLong = descriptionLong
 
-    // Event Start and End Dates and Times
-    if (!data.filteredEvents.endDates[eventEndDate]) {
-      data.filteredEvents.endDates[eventEndDate] = []
-    }
-    data.filteredEvents.endDates[eventEndDate].push(index)
-    newEvent.endDate = eventEndDate
-
-    if (!data.filteredEvents.endTimes[eventEndTime]) {
-      data.filteredEvents.endTimes[eventEndTime] = []
-    }
-    data.filteredEvents.endTimes[eventEndTime].push(index)
-    newEvent.endTime = eventEndTime
-
-    if (!data.filteredEvents.startDates[eventStartDate]) {
-      data.filteredEvents.startDates[eventStartDate] = []
-    }
-    data.filteredEvents.startDates[eventStartDate].push(index)
-    newEvent.startDate = eventStartDate
-
-    if (!data.filteredEvents.startTimes[eventStartTime]) {
-      data.filteredEvents.startTimes[eventStartTime] = []
-    }
-    data.filteredEvents.startTimes[eventStartTime].push(index)
-    newEvent.startTime = eventStartTime
-
     // Event Type
     if (!data.filteredEvents.eventTypes[eventType]) {
       data.filteredEvents.eventTypes[eventType] = []
     }
-    data.filteredEvents.eventTypes[eventType].push(index)
+    data.filteredEvents.eventTypes[eventType].push(newId)
     newEvent.eventType = eventType
 
     // Experience Type
     if (!data.filteredEvents.experienceType[experienceReq]) {
       data.filteredEvents.experienceType[experienceReq] = [];
     }
-    data.filteredEvents.experienceType[experienceReq].push(index)
+    data.filteredEvents.experienceType[experienceReq].push(newId)
     newEvent.experienceType = experienceReq
 
     // Game Id
     newEvent.gameId = gameId
 
+
     // Game System
     if (gameSystem) {
-      if (!data.filteredEvents.gameSystems[gameSystem]) {
-        data.filteredEvents.gameSystems[gameSystem] = []
+      var gameSystemLabel = gameSystem;
+
+      gameSystemLabel = gameSystem.toString().replace(/:/g,'');
+      // gameSystemLabel = gameSystemLabel.toUpperCase();
+
+      if (!data.filteredEvents.gameSystems[gameSystemLabel]) {
+        data.filteredEvents.gameSystems[gameSystemLabel] = []
       }
-      data.filteredEvents.gameSystems[gameSystem].push(index)
+      data.filteredEvents.gameSystems[gameSystemLabel].push(newId)
       newEvent.gameSystem = gameSystem
     }
 
@@ -187,22 +232,22 @@ export default function cleanData (eventData: RawEvent[]) {
       if (!data.filteredEvents.groups[group]) {
         data.filteredEvents.groups[group] = []
       }
-      data.filteredEvents.groups[group].push(index)
+      data.filteredEvents.groups[group].push(newId)
       newEvent.group = group
     }
 
     // Location
     if (location) {
-      if (!data.filteredEvents.locations[location]) {
-        data.filteredEvents.locations[location] = []
+      if (!data.filteredEvents.locations[location.toUpperCase()]) {
+        data.filteredEvents.locations[location.toUpperCase()] = []
       }
-      data.filteredEvents.locations[location].push(index)
+      data.filteredEvents.locations[location.toUpperCase()].push(newId)
       newEvent.location = location
     }
 
     // Materials Required
     if (materials) {
-      data.filteredEvents.materialsRequired.push(index)
+      data.filteredEvents.materialsRequired.push(newId)
       newEvent.materials = materials
     }
 
@@ -226,7 +271,7 @@ export default function cleanData (eventData: RawEvent[]) {
 
     // Tickets
     if (!ticketsAvailable) {
-      data.filteredEvents.noTickets.push(index)
+      data.filteredEvents.noTickets.push(newId)
     }
     newEvent.ticketsAvailable = ticketsAvailable
 
@@ -235,7 +280,7 @@ export default function cleanData (eventData: RawEvent[]) {
 
     // Tournament 
     if (tournament) {
-      data.filteredEvents.tournaments.push(index)
+      data.filteredEvents.tournaments.push(newId)
     }
     newEvent.tournament = tournament
 
@@ -251,5 +296,5 @@ export default function cleanData (eventData: RawEvent[]) {
     data.eventData.push(newEvent)
   })
 
-  return data
+  return data;
 }
