+++ frontend/pages/_app.tsx
@@ -6,7 +6,6 @@ import type { AppProps } from 'next/app'
 import Head from 'next/head'
 import { Analytics } from '@vercel/analytics/react'
 
-import getData from "@/helpers/getData"
 import parseData from "@/helpers/parseData"
 
 export const { eventData, filteredEvents } = parseData()
